backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
opencv-python==4.8.1.78
scikit-learn==1.3.2
numpy==1.24.4
pandas==1.5.3
pytesseract==0.3.10
python-barcode==0.15.1
pyzbar==0.1.9
tensorflow==2.15.0
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
easyocr==1.7.0
joblib==1.3.2
requests==2.31.0
python-dotenv==1.0.0
cors==1.0.1
aiofiles==23.2.1